Clean up Examples (#695)
Restructure examples to use ViewStorage (except counter and variants that use SimpleStorage)

* Removed duplicate examples (i.e. 'x' and 'x2' or 'x-graphql')

* Fixed unit tests to work with reduced examples

* Updated end_to_end tests with new examples

* Formatting and clippy

* Use sanitizer for worker tests

* Final nits on the example refactoring

* fix linter

+++ examples/counter/src/state.rs
@@ -6,5 +6,5 @@ use serde::{Deserialize, Serialize};
 /// The application state.
 #[derive(Clone, Copy, Debug, Default, Deserialize, Serialize)]
 pub struct Counter {
-    pub value: u128,
+    pub value: u64,
 }
